% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{all_possible_origins}
\alias{all_possible_origins}
\title{All possible 3-digit origins}
\format{An object of class \code{character} of length 1000.}
\usage{
all_possible_origins
}
\description{
All possible 3-digit origins
}
\keyword{datasets}
