% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{detail_definitions}
\alias{detail_definitions}
\title{Details}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.}
\usage{
detail_definitions
}
\description{
Details
}
\keyword{datasets}
