% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all.R
\name{fetch_all}
\alias{fetch_all}
\title{Grab all the origin-destination pairs.}
\usage{
fetch_all(origins = all_possible_origins, sleep_time = 1, verbose = TRUE,
  ...)
}
\arguments{
\item{origins}{A vector of origin zips.}

\item{sleep_time}{How long to sleep in between requests, plus or minus \code{runif(1)} second.}

\item{verbose}{Message what's going on?}

\item{...}{Other arguments}
}
\value{
A tibble with origin zip and destination zips (in ranges or unspooled) and the USPS zones the origin-destination pair corresponds to.
}
\description{
For a given 3-digit origin zip code, grab all destination zips and their corresponding zones.
}
\details{

}
\examples{
\dontrun{

fetch_all(sample(all_possible_origins, 4))
}

}
