% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_five_digit.R
\name{fetch_five_digit}
\alias{fetch_five_digit}
\title{Fetch zones for a 5-digit origin origin-destination pair}
\usage{
fetch_five_digit(origin_zip, destination_zip, ...)
}
\arguments{
\item{origin_zip}{A single origin zip as 5-digit character.}

\item{destination_zip}{Optional destination zip as 5-digit character.}

\item{...}{Other arguments}
}
\value{
A tibble with origin zip and destination zips (in ranges or unspooled) and the USPS zones the origin-destination pair corresponds to.
}
\description{
For a given 5-digit origin and destination zip code pair, display the zone number and full response.
}
\details{
Displays the result of a query to the ["Get Zone for ZIP Code Pair"](https://postcalc.usps.com/DomesticZoneChart/) tab.
}
\examples{
\dontrun{
fetch_five_digit("90210", "20500")
}

}
