% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mail_flat_rate.R
\name{fetch_mail_flat_rate}
\alias{fetch_mail_flat_rate}
\title{Get postage options for a flat-rate envelope or box}
\usage{
fetch_mail_flat_rate(origin_zip = NULL, destination_zip = NULL,
  shipping_date = "today", shipping_time = "now", type = c("envelope",
  "box"), ground_transportation_needed = FALSE, live_animals = FALSE,
  day_old_poultry = FALSE, hazardous_materials = FALSE,
  show_details = FALSE, n_tries = 3, verbose = TRUE, ...)
}
\arguments{
\item{origin_zip}{5-digit origin zip code.}

\item{destination_zip}{5-digit destination zip code.}

\item{shipping_date}{Date you plan to ship the package on in "MM-DD-YYYY" format as character, or "today".}

\item{shipping_time}{Time of day you plan to ship in 24-hour "HH:MM" format as character, or "now".}

\item{type}{One of: "box", "envelope".}

\item{ground_transportation_needed}{Does the package need to be transported by ground?}

\item{live_animals}{Boolean: does this contain live animals? See: \url{https://pe.usps.com/text/pub52/pub52c5_003.htm}}

\item{day_old_poultry}{Boolean: does this contain day-old poultry? See: \url{https://pe.usps.com/text/pub52/pub52c5_008.htm#ep184002}}

\item{hazardous_materials}{Boolean: does this contain any hazardous materials? See: \url{https://pe.usps.com/text/pub52/pub52c3_001.htm}}

\item{show_details}{Non-essential details of the response are hidden by default. Show them by setting this to TRUE.}

\item{n_tries}{How many times to try the API if at first we don't succeed?}

\item{verbose}{Should messages, (e.g. shipping date time be dispalyed if the defaults "today" and "now" are chosen) be messageed?}

\item{...}{Other arguments.}
}
\value{
A tibble with information for different postage options, including price and box/envelope dimensions.
}
\description{
Get postage options for a flat-rate envelope or box
}
\details{
Supply the required information about the package and receive a tibble. Displays the result of a query to the ["Postage Price Calculator"](https://postcalc.usps.com/Calculator/). For non-flat-rate packages, use \code{\link{fetch_mail_package}}.
}
\examples{
\dontrun{

fetch_mail_flat_rate(origin_zip = "60647",
         destination_zip = "11238", type = "envelope")
}

}
