% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mail_package.R
\name{fetch_mail_package}
\alias{fetch_mail_package}
\title{Get postage options for a package}
\usage{
fetch_mail_package(origin_zip = NULL, destination_zip = NULL,
  shipping_date = "today", shipping_time = "now",
  ground_transportation_needed = FALSE, live_animals = FALSE,
  day_old_poultry = FALSE, hazardous_materials = FALSE, pounds = 0,
  ounces = 0, length = 0, height = 0, width = 0, girth = 0,
  shape = c("rectangular", "nonrectangular"), show_details = FALSE,
  n_tries = 3, verbose = TRUE, ...)
}
\arguments{
\item{origin_zip}{5-digit origin zip code.}

\item{destination_zip}{5-digit destination zip code.}

\item{shipping_date}{Date you plan to ship the package on in "MM-DD-YYYY" format as character, or "today".}

\item{shipping_time}{Time of day you plan to ship in 24-hour "HH:MM" format as character, or "now".}

\item{ground_transportation_needed}{Does the package need to be transported by ground?}

\item{live_animals}{Boolean: does this contain live animals?}

\item{day_old_poultry}{Boolean: does this contain day-old poultry?}

\item{hazardous_materials}{Boolean: does this contain any hazardous materials? See: \url{https://pe.usps.com/text/pub52/pub52c3_001.htm}}

\item{pounds}{Number of pounds the package weighs.}

\item{ounces}{Number of ounces the package weighs.}

\item{length}{Length of the package. This is the longest dimension.}

\item{height}{Height of the package.}

\item{width}{Width of the package.}

\item{girth}{Girth of the package, required if \code{shape} is "nonrectangular". This is the distance around the thickest part.}

\item{shape}{Shape of the package: "rectangular" or "nonrectangular". "nonrectangular" reqires a non-null \code{girth} value.}

\item{show_details}{Non-essential details of the response are hidden by default. Show them by setting this to TRUE.}

\item{n_tries}{How many times to try the API if at first we don't succeed?}

\item{verbose}{Should messages, (e.g. shipping date time be dispalyed if the defaults "today" and "now" are chosen) be messageed?}

\item{...}{Other arguments.}
}
\value{
A tibble with information for different postage options.
}
\description{
Supply some information about the package.
}
\details{
Displays the result of a query to the ["Postage Price Calculator"](https://postcalc.usps.com/Calculator/). For flat rate envelopes or boxes, use \code{\link{fetch_mail_flat_rate}}.
}
\examples{
\dontrun{

fetch_mail_package(origin_zip = "60647",
         destination_zip = "11238",
         pounds = 15, shape = "rectangular")
}

}
