% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_zone.R
\name{grab_zone_from_origin}
\alias{grab_zone_from_origin}
\title{Grab destination zones}
\usage{
grab_zone_from_origin(origin_zip, as_range = FALSE, show_modifiers = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{origin_zip}{A single origin zip, numeric or character}

\item{as_range}{Do you want zones corresponding to a range of destination zips or a full listing of them?}

\item{show_modifiers}{Should columns pertaining to the modifiers * and + be retained?}

\item{verbose}{Message what's going on?}

\item{...}{Other arguments}
}
\value{
A tibble with origin zip and destination zips (in ranges or unspooled) and the USPS zones the origin-destination pair corresponds to.
Validity attribute lets you know whether the origin zip code is in use (see also https://en.wikipedia.org/wiki/List_of_ZIP_code_prefixes)
}
\description{
For a given 3-digit origin zip code, grab all destination zips and their corresponding zones.
}
\details{
https://postcalc.usps.com/
}
