% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{has_mail}
\alias{has_mail}
\title{Test if \code{mail} is installed on the host platform
See \url{https://www.binarytides.com/linux-mail-command-examples/}}
\usage{
has_mail()
}
\value{
logical, TRUE if present
}
\description{
Test if \code{mail} is installed on the host platform
See \url{https://www.binarytides.com/linux-mail-command-examples/}
}
\seealso{
Other mail: 
\code{\link{has_mutt}()},
\code{\link{has_nail}()},
\code{\link{muttmail}()},
\code{\link{nailmail}()},
\code{\link{sendmail}()}
}
\concept{mail}
