% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{sendmail}
\alias{sendmail}
\title{Send an email from R using the underlying mail services.}
\usage{
sendmail(..., app = c("mail", "mutt", "nail")[1])
}
\arguments{
\item{...}{arguments for the mailer of choice.}

\item{app}{the name of the mail engine to use, currently 'mail' (default), nail' and 'mutt' are supported}
}
\value{
value returned by the system-level mail program, non-zero for failure.
}
\description{
A simple wrapper around the platforms email engine.  Currently \code{mutt},
\code{nail} and \code{mail} are supported, but the host must have these installed.
}
\seealso{
Other mail: 
\code{\link{has_mail}()},
\code{\link{has_mutt}()},
\code{\link{has_nail}()},
\code{\link{muttmail}()},
\code{\link{nailmail}()}
}
\concept{mail}
