% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{zone_detail_definitions}
\alias{zone_detail_definitions}
\title{Details}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.}
\usage{
zone_detail_definitions
}
\description{
Details
}
\keyword{datasets}
