% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_zones_five_digit.R
\name{fetch_five_digit}
\alias{fetch_five_digit}
\title{Fetch zones for a 5-digit origin origin-destination pair}
\usage{
fetch_five_digit(origin_zip, destination_zip, n_tries = 3,
  show_details = FALSE, ...)
}
\arguments{
\item{origin_zip}{A single origin zip as 5-digit character.}

\item{destination_zip}{Required destination zip as 5-digit character.}

\item{n_tries}{How many times to try the API if at first we don't succeed?}

\item{show_details}{Extract extra stuff from the response?}

\item{...}{Other arguments}

\item{verbose}{Message what's going on?}
}
\value{
A tibble with origin zip and destination zips (in ranges or unspooled) and the USPS zones the origin-destination pair corresponds to.
}
\description{
For a given 5-digit origin and destination zip code pair, display the zone number and full response.
}
\details{
Displays the result of a query to the ["Get Zone for ZIP Code Pair"](https://postcalc.usps.com/DomesticZoneChart/) tab.

If you want all destinations for a given origin, use \code{\link{fetch_zones}} with the first 3 digits of the origin; there you don't need to supply a destination.
}
\examples{
\dontrun{
fetch_five_digit("90210", "20500")

fetch_five_digit("40360", "09756", show_details = TRUE)

purrr::map2_dfr(c("11238", "60647", "80205"),
                c("98109", "02210", "94707"),
      fetch_five_digit)
}

}
