% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mail_flat_rate.R
\name{fetch_mail_flat_rate}
\alias{fetch_mail_flat_rate}
\title{Get postage options for a package}
\usage{
fetch_mail_flat_rate(origin_zip = NULL, destination_zip = NULL,
  shipping_date = "today", shipping_time = "now", type = "box",
  ground_transportation_needed = FALSE, show_details = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{origin_zip}{A single origin zip as character. If > 3 digits and contains leading zeros, make sure to supply as character.}

\item{destination_zip}{Optional destination zip. If not included, returns all possible destinations for the origin provided. If > 3 digits and contains leading zeros, make sure to supply as character.}

\item{shipping_date}{Date you plan to ship the package on in "MM-DD-YYY" format as character, or "today".}

\item{shipping_time}{Time of day you plan to ship in "HH:MM" form, or "now".}

\item{type}{One of: "box", "envelope".}

\item{ground_transportation_needed}{Does the package need to be transported by ground?}

\item{show_details}{Extra details?}

\item{verbose}{More messages}

\item{...}{Other args}
}
\value{
A tibble with information for different postage options.
}
\description{
Supply some information about the package.
}
\details{
Displays the result of a query to the ["Postage Price Calculator"](https://postcalc.usps.com/Calculator/).
}
\examples{
\dontrun{

get_mail(origin_zip = 60647,
         destination_zip = 11238,
         pounds = 15)
}

}
