% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mail_package.R
\name{fetch_mail_package}
\alias{fetch_mail_package}
\title{Get postage options for a package}
\usage{
fetch_mail_package(origin_zip = NULL, destination_zip = NULL,
  shipping_date = "today", shipping_time = "now",
  ground_transportation_needed = FALSE, live_animals = FALSE,
  day_old_poultry = FALSE, hazardous_materials = FALSE, pounds = 0,
  ounces = 0, length = 0, height = 0, width = 0, girth = 0,
  shape = "rectangular", show_details = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{origin_zip}{A single origin zip as character. If > 3 digits and contains leading zeros, make sure to supply as character.}

\item{destination_zip}{Optional destination zip. If not included, returns all possible destinations for the origin provided. If > 3 digits and contains leading zeros, make sure to supply as character.}

\item{shipping_date}{Date you plan to ship the package on in "MM-DD-YYY" format as character, or "today".}

\item{shipping_time}{Time of day you plan to ship in "HH:MM" form, or "now".}

\item{ground_transportation_needed}{Does the package need to be transported by ground?}

\item{live_animals}{Boolean: does this contain live animals?}

\item{day_old_poultry}{Boolean: does this contain day-old poultry?}

\item{hazardous_materials}{Boolean: does this contain any hazardous materials? See: \url{https://pe.usps.com/text/pub52/pub52c3_001.htm}}

\item{pounds}{Number of pounds the package weighs.}

\item{ounces}{Number of ounces the package weighs.}

\item{length}{Length of the package. This is the longest dimension.}

\item{height}{Height of the package.}

\item{width}{Width of the package.}

\item{girth}{Girth of the package, required if \code{shape} is "nonrectangular". This is the distance around the thickest part.}

\item{shape}{Shape of the package: "rectangular" or "nonrectangular". "Nonrectangular" reqires a non-null \code{girth} value.}

\item{show_details}{Extra details?}

\item{verbose}{More messages}

\item{...}{Other args}
}
\value{
A tibble with information for different postage options.
}
\description{
Supply some information about the package.
}
\details{
Displays the result of a query to the ["Postage Price Calculator"](https://postcalc.usps.com/Calculator/).
}
\examples{
\dontrun{

fetch_mail_package(origin_zip = "60647",
         destination_zip = "11238",
         pounds = 15)
}

}
