% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_zones.R
\name{fetch_zones}
\alias{fetch_zones}
\title{Fetch zones for a 3-digit origin zip or an origin-destination pair}
\usage{
fetch_zones(origin_zip = NULL, destination_zip = NULL, as_range = FALSE,
  show_modifiers = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{origin_zip}{A single origin zip, numeric or character}

\item{destination_zip}{Optional destination zip. If not included, returns all possible desinations for the origin provided.}

\item{as_range}{Do you want zones corresponding to a range of destination zips or a full listing of them?}

\item{show_modifiers}{Should columns pertaining to the modifiers * and + be retained?}

\item{verbose}{Message what's going on?}

\item{...}{Other arguments}
}
\value{
A tibble with origin zip and destination zips (in ranges or unspooled) and the USPS zones the origin-destination pair corresponds to.
Validity attribute lets you know whether the origin zip code is in use (see also \url{https://en.wikipedia.org/wiki/List_of_ZIP_code_prefixes})
}
\description{
For a given 3-digit origin zip code, grab all destination zips and their corresponding zones.
}
\details{
\url{https://postcalc.usps.com/}
}
\examples{
\dontrun{

a_zip <- fetch_zones(123)
nrow(a_zip)

fetch_zones(123, 456, show_modifiers = TRUE)

(double_oh_seven <- fetch_zones("007", as_range = TRUE))
attr(double_oh_seven, "validity")}

}
