% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usps-package.R
\docType{package}
\name{usps}
\alias{usps}
\alias{usps-package}
\title{Fetch mails and zones from USPS}
\description{
Calculate shipping rates and times for packages and get the USPS zones corresponding to 3-digit and 5-digit zip code pairs.
}
\details{
Contributors:
\itemize{
 \item Amanda Dobbyn
}

To get postage information, use \code{fetch_mail}.

To get zones, use \code{\link{get_zone_three_digit}} or \code{\link{get_zone_five_digit}}

say('hello world!')
}
