% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zips_zones}
\alias{zips_zones}
\title{Zips and Zones}
\format{A data frame with 3,804,494 rows and 6 variables:
\describe{
  \item{origin_zip}{price, in US dollars}
  \item{dest_zip_start}{weight of the diamond, in carats}
  \item{dest_zip_end}{weight of the diamond, in carats}
  \item{zone}{weight of the diamond, in carats}
}}
\usage{
zips_zones
}
\description{
All 3-digit zips and zones. The result of running fetch_all() with \code{as_range = TRUE}.
}
\keyword{datasets}
